+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-editor/case-edit/case-edit.component.spec.ts
@@ -19,6 +19,7 @@ import { CaseNotifier, WorkAllocationService } from '../services';
 import { WizardFactoryService } from '../services/wizard-factory.service';
 import { ValidPageListCaseFieldsService } from '../services/valid-page-list-caseFields.service';
 import { CaseEditComponent } from './case-edit.component';
+import { AbstractAppConfig } from '../../../../app.config';
 import createSpyObj = jasmine.createSpyObj;
 
 describe('CaseEditComponent', () => {
@@ -214,6 +215,7 @@ describe('CaseEditComponent', () => {
   let mockSessionStorageService: jasmine.SpyObj<SessionStorageService>;
   let mockWorkAllocationService: jasmine.SpyObj<WorkAllocationService>;
   let mockAlertService: jasmine.SpyObj<AlertService>;
+  let mockabstractConfig: jasmine.SpyObj<AbstractAppConfig>;
   const validPageListCaseFieldsService = new ValidPageListCaseFieldsService(fieldsUtils);
 
   describe('profile available in route', () => {
@@ -274,7 +276,8 @@ describe('CaseEditComponent', () => {
       ]);
       mockSessionStorageService = createSpyObj<SessionStorageService>('SessionStorageService', ['getItem', 'removeItem', 'setItem']);
       mockWorkAllocationService = createSpyObj<WorkAllocationService>('WorkAllocationService', ['assignAndCompleteTask', 'completeTask']);
-      mockAlertService = createSpyObj<AlertService>('WorkAllocationService', ['error', 'setPreserveAlerts']);
+      mockAlertService = createSpyObj<AlertService>('AlertService', ['error', 'setPreserveAlerts']);
+      mockabstractConfig = createSpyObj<AbstractAppConfig>('AbstractAppConfig', ['logMessage']);
       spyOn(validPageListCaseFieldsService, 'deleteNonValidatedFields');
       spyOn(validPageListCaseFieldsService, 'validPageListCaseFields');
 
@@ -330,6 +333,7 @@ describe('CaseEditComponent', () => {
             WindowService,
             { provide: LoadingService, loadingServiceMock },
             { provide: ValidPageListCaseFieldsService, useValue: validPageListCaseFieldsService},
+            { provide: AbstractAppConfig, useValue: mockabstractConfig },
           ]
         })
         .compileComponents();
@@ -1300,8 +1304,8 @@ describe('CaseEditComponent', () => {
           form: component.form,
           submit: mockClass.submit,
         });
-
-        expect(mockWorkAllocationService.assignAndCompleteTask).toHaveBeenCalledWith('12345');
+        expect(mockabstractConfig.logMessage).toHaveBeenCalledWith('postCompleteTaskIfRequired with assignNeeded: taskId 12345 and event name Test Trigger');
+        expect(mockWorkAllocationService.assignAndCompleteTask).toHaveBeenCalledWith('12345', component.eventTrigger.name);
       });
 
       it('should submit the case and complete task for an event submission', () => {
@@ -1336,8 +1340,8 @@ describe('CaseEditComponent', () => {
           form: component.form,
           submit: mockClass.submit,
         });
-
-        expect(mockWorkAllocationService.completeTask).toHaveBeenCalledWith('12345');
+        expect(mockabstractConfig.logMessage).toHaveBeenCalledWith('postCompleteTaskIfRequired: taskId 12345 and event name Test Trigger');
+        expect(mockWorkAllocationService.completeTask).toHaveBeenCalledWith('12345', component.eventTrigger.name);
       });
 
       it('should NOT submit the case due to error', () => {
@@ -1560,6 +1564,7 @@ describe('CaseEditComponent', () => {
             SessionStorageService,
             WindowService,
             { provide: ValidPageListCaseFieldsService, useValue: validPageListCaseFieldsService},
+            { provide: AbstractAppConfig, useValue: mockabstractConfig },
           ]
         })
         .compileComponents();